# docs/swagger.yaml
basePath: /
definitions:
  models.CreateArticleRequest:
    properties:
      content:
        minLength: 1
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
  models.CreateCommentRequest:
    properties:
      content:
        minLength: 1
        type: string
    required:
    - content
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.UpdateArticleRequest:
    properties:
      content:
        minLength: 1
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
info:
  contact: {}
  description: 微服务 code  by 田发亮
  title: 博客系统
  version: 0.0.1
paths:
  /api/articles:
    get:
      consumes:
      - application/json
      description: GetArticles
      parameters:
      - default: Bearer
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: name,helloWorld
          schema:
            type: string
    post:
      consumes:
      - application/json
      description: CreateArticle
      parameters:
      - default: Bearer
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: name,helloWorld
          schema:
            type: string
  /api/articles/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteArticle
      parameters:
      - default: Bearer
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: name,helloWorld
          schema:
            type: string
    get:
      consumes:
      - application/json
      description: GetArticle
      parameters:
      - default: Bearer
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: name,helloWorld
          schema:
            type: string
    put:
      consumes:
      - application/json
      description: UpdateArticle
      parameters:
      - default: Bearer
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateArticleRequest
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: name,helloWorld
          schema:
            type: string
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: name,helloWorld
          schema:
            type: string
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register
      parameters:
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: username,helloWorld
          schema:
            type: string
  /api/comments/{articleId}:
    get:
      consumes:
      - application/json
      description: GetArticleComments
      parameters:
      - default: Bearer
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: articleId
        in: path
        name: articleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: name,helloWorld
          schema:
            type: string
    post:
      consumes:
      - application/json
      description: CreateComment
      parameters:
      - default: Bearer
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: articleId
        in: path
        name: articleId
        required: true
        type: integer
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: name,helloWorld
          schema:
            type: string
swagger: "2.0"